+++ include/active_record/row.h
@@ -30,7 +30,7 @@ class Row {
   Type::Type get_type(const string& name);
   bool is_type(const string& name, Type::Type type);
   string get_text(const string& name);
-  int get_integer(const string& name);
+  int64 get_integer(const string& name);
   double get_floating_point(const string& name);
   Date get_date(const string& name);
   string to_string() const;
